Add files via upload
This is for Leetcode problem No.5 Longest Palindromic Substring. This is my initial "brute force" solution, I used two for-loop to get all the substring and use a Boolean function to determine whether it is palindromic. It works well, but when it comes to very long string like string with length 900 it will exceed the time limit.